pyproject.toml
[project]
name = "streamflow-percentiles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataretrieval>=1.1.2",
    "folium>=0.20.0",
    "geopandas>=1.1.2",
    "hyswap>=1.0.0",
    "ipympl>=0.10.0",
    "mapclassify>=2.10.0",
]
